Fix warehouse deserialization: Volume field is double, not int
The warehouse endpoint returns volume as a floating-point number

(e.g. 364.0) while the asset endpoint returns it as an integer (50).

+++ OE2EmpireTracker.Common/Models/GameApiAssetResponse.cs
@@ -154,7 +154,7 @@ namespace OE2EmpireTracker.Client
         /// Gets or sets the volume of the item.
         /// </summary>
         [JsonProperty("volume")]
-        public int? Volume { get; set; }
+        public double? Volume { get; set; }
 
         /// <summary>
         /// Gets or sets the item properties (modifications, stats).
